tests/wpt/web-platform-tests/viewport/viewport-type.html

<!doctype html>
<html>
    <head>
        <title>Viewport: window.view type</title>
        <meta charset="utf-8">
        <meta name="viewport" content="width=device-width">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
    <h1>Viewport: window.view type</h1>
    <h4>Test Description: This test checks that window.view is an object of type VisualViewport.</h4>
    <script>
        test(function() {
              assert_equals(typeof(window.view), "object");
            }, "window.view is an object");

        test(function() {
              assert_equals(window.view.toString(), "[object VisualViewport]");
            }, "window.view has type `VisualViewport`");

        addEventListener("load", function() {
            document.getElementById("viewtype-log").innerText = typeof(window.view);
        });
    </script>
    <div id="complete-notice">
        <p>window.view is of type: <span id="viewtype-log"></span>.</p>
    </div>
    <div id="log"></div>
    </body>
</html>
