LayoutTests/imported/w3c/web-platform-tests/viewport/viewport-type-expected.txt

Viewport: window.visualViewport type

Test Description: This test checks that window.visualViewport is an object of type VisualViewport.

window.visualViewport is of type: undefined.


FAIL window.visualViewport is an object assert_equals: expected "object" but got "undefined"
FAIL window.visualViewport has type `VisualViewport` undefined is not an object (evaluating 'window.visualViewport.toString')
